[FEAT] #14 - HelloWorldActivity 삭제 + 내용은 ShopppingFragment로
