+++ feature/around/src/main/java/com/swmarastro/mykkumi/feature/around/AroundFragment.kt
@@ -11,7 +11,7 @@ class AroundFragment : BaseFragment<FragmentAroundBinding>(R.layout.fragment_aro
 
     private val viewModel by viewModels<AroundViewModel>()
 
-    override fun initView() {
+    override suspend fun initView() {
         bind {
             vm = viewModel
         }
