--- feature/home/src/main/res/layout/activity_hello_world.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.swmarastro.mykkumi.feature.home.hello.HelloWorldActivity">
-    
-    <data>
-        <variable
-            name="vm"
-            type="com.swmarastro.mykkumi.feature.home.hello.HelloWorldViewModel" />
-    </data>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/white">
-
-        <TextView
-            android:id="@+id/text_helloworld"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/black"
-            android:textSize="14dp"
-            android:layout_centerInParent="true"/>
-
-    </RelativeLayout>
-
-</layout>
