+++ feature/shopping/src/main/res/layout/fragment_shopping.xml
@@ -15,9 +15,18 @@
         android:layout_height="match_parent">
 
         <TextView
+            android:id="@+id/text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="쇼핑"/>
+
+        <TextView
+            android:id="@+id/text_helloworld"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textColor="@color/black"
+            android:textSize="14dp"
+            android:layout_below="@id/text"/>
 
     </RelativeLayout>
